PW-1: Adding multi-partner handling in metering v4

+++ metering/views.py
@@ -1,6 +1,7 @@
 #Copyright 2015 Phoenix Bioinformatics Corporation. All rights reserved.
 
 from metering.models import ipAddr, limits
+from django.db.models import Max
 from metering.serializers import ipSerializer, limitSerializer
 from rest_framework import generics
 import json
@@ -10,18 +11,23 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import redirect
-
+from partner.models import Partner
 ###############################################For IP##############################################
 #List view of all IP counts
 class ipList(APIView):
     def get(self, request, format=None):
-        ips = ipAddr.objects.all()
-        partnerId = request.QUERY_PARAMS.get('partnerId', None)
-        ips = ips.filter(partner=partnerId)
+        ips = Partner.filters(self, ipAddr.objects.all(), "partner")
         serializer = ipSerializer(ips, many=True)
         return Response(serializer.data, status.HTTP_200_OK)
 
     def post(self, request, format=None):
+        if request.GET.get('partnerId', 'None') is None:
+          return Response(status=status.HTTP_400_BAD_REQUEST)
+        if request.GET.get('partnerId')!=request.data['partner']:
+          raise Http404
+        checkIp = Partner.filters(self, ipAddr.objects.all(), "partner").filter(ip=request.data['ip'])
+        if checkIp:
+          return Response(status=status.HTTP_400_BAD_REQUEST)
         serializer = ipSerializer(data=request.data)
         if serializer.is_valid():
           serializer.save()
@@ -29,30 +35,34 @@ class ipList(APIView):
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, format=None):
-        ips = ipAddr.objects.all()
-        partnerId = request.QUERY_PARAMS.get('partnerId', None)
-        ips = ips.filter(partner=partnerId)
-        ips.delete()
+        ips = Partner.filters(self, ipAddr.objects.all(), "partner")
+        if ips:
+          ips.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 #Detail view for each IP
 class ipDetail(APIView):
     def get(self, request, pk, format=None):
-        partnerId = request.QUERY_PARAMS.get('partnerId', None)
-        snippet = get_object(pk, partnerId)
-        serializer = ipSerializer(snippet, many=True)
+        snippet = Partner.filters(self, ipAddr.objects.all(), "partner")
+        if snippet:
+          snippet = snippet.get(ip=pk)
+          serializer = ipSerializer(snippet)
+        else:
+          serializer = ipSerializer(snippet, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def delete(self, request, pk, format=None):
-        partnerId = request.QUERY_PARAMS.get('partnerId', None)
-        snippet = get_object(pk, partnerId)
-        snippet.delete()
+        snippet = Partner.filters(self, ipAddr.objects.all(), "partner")
+        if snippet:
+          snippet = snippet.get(ip=pk)
+          snippet.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     def put(self, request, pk, format=None):
-        partnerId = request.QUERY_PARAMS.get('partnerId', None)
-        snippet = get_object(pk, partnerId)
+        snippet = Partner.filters(self, ipAddr.objects.all(), "partner")
+        if snippet:
+          snippet.get(ip=pk)
         serializer = ipSerializer(snippet, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -60,93 +70,75 @@ class ipDetail(APIView):
         return Response(status=status.HTTP_400_BAD_REQUEST)
         
 #Increment request for an IP
-def increment(request, ip, format=None):
-    partnerId = request.QUERY_PARAMS.get('partnerId', None)
-    currIp = get_object(ip, partnerId)
-    if (currIp.count < get_limit("MeteringLimit")):
-        currIp.count += 1
-        currIp.save()
+class increment(APIView):
+  def get(self, request, ip, format=None):
+    currIp = Partner.filters(self, ipAddr.objects.all(), "partner")
+    if currIp:
+      currIp = currIp.get(ip=ip)
+    #TODO: if (currIp.count < get_limit("MeteringLimit")):
+      currIp.count += 1
+      currIp.save()
+    #TODO:
     return redirect('/meters/ip/'+ip+'/')
 
 ############################################For Limits#############################################
 #Return limit status of IP
-def check_limit(request, ip, format=None):
-    currIp = get_object(ip)
-    if (currIp.count == get_limit("WarningLimit").val):
-        ret = {'status': "Warning"}
-    elif (currIp.count < get_limit("MeteringLimit").val):
-        ret = {'status': "OK"}
-    else:
+class check_limit(APIView):
+  def get(self, request, ip, format=None):
+    currIp = Partner.filters(self, ipAddr.objects.all(), "partner")
+    limitVal = Partner.filters(self, limits.objects.all(), "partner")
+    if currIp:
+      currIp = currIp.get(ip=ip)
+    if not limitVal:
+      return Response(status=status.HTTP_400_BAD_REQUEST)
+    
+    if (currIp.count >= limitVal.aggregate(Max('val'))['val__max']):
         ret = {'status': "Block"}
+    elif (currIp.count in limitVal.values_list('val', flat=True)):
+        ret = {'status': "Warning"}
+    else:
+        ret = {'status': "OK"}
     return HttpResponse(json.dumps(ret), content_type="application/json", status=200)
 
 #Detail view/update for warning limit
 class warningLimit(APIView):
     def get(self, request, format=None):
-        limit = get_limit("WarningLimit")
-        serializer = limitSerializer(limit)
+        limit = Partner.filters(self, limits.objects.all(), "partner")
+        serializer = limitSerializer(limit, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    def put(self, request, format=None):
-        limit = get_create_limit("WarningLimit")
-        limit.val = request.data['limit']
-        limit.save(using='mySQLTest')
-        serializer = limitSerializer(limit)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+    def post(self, request, format=None):
+        if request.GET.get('partnerId', 'None') is None:
+          return Response(status=status.HTTP_400_BAD_REQUEST)
+        if request.GET.get('partnerId')!=request.data['partner']:
+          raise Http404
+        checkLimit = Partner.filters(self, limits.objects.all(), "partner").filter(val=request.data['val'])
+        if checkILimit:
+          return Response(status=status.HTTP_400_BAD_REQUEST)
+        serializer = limitSerializer(data=request.data)
+        if serializer.is_valid():
+          serializer.save()
+          return Response(status=status.HTTP_201_CREATED)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
 #Detail view/update for metering limit
 class meteringLimit(APIView):
     def get(self, request, format=None):
-        limit = get_limit("MeteringLimit")
-        serializer = limitSerializer(limit)
+        limit = Partner.filters(self, limits.objects.all(), "partner")
+        if limit:
+          limit = limit.order_by('val').last()
+          serializer = limitSerializer(limit)
+        else:
+          serializer = limitSerializer(limit, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, format=None):
-        limit = get_create_limit("MeteringLimit")
-        limit.val = request.data['limit']
-        limit.save(using='mySQLTest')
-        serializer = limitSerializer(limit)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-
+        snippet = Partner.filters(self, limits.objects.all(), "partner")
+        if snippet:
+          snippet = snippet.order_by('val').last()
+        serializer = ipSerializer(snippet, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 ###########################################Auxillary functions#####################################
-def get_object(pk, partner):
-    try:
-        ips = ipAddr.objects.all()
-        if partner is not None:
-            ips = ipAddr.objects.filter(partner=partner)
-        return ips.filter(ip=pk)
-    except ipAddr.DoesNotExist:
-        #TODO: Put status message for debug
-        raise Http404
-
-def get_limit(lm):
-    try: 
-        return limits.objects.using('mySQLTest').get(name=lm)
-    except:
-        #TODO: Put status message for debug
-        raise Http404
-
-def get_create_limit(lm):
-    try:
-        return limits.objects.get(name=lm)
-    except:
-        u = limits(name=lm, val=0)
-        u.save()
-        return u
-        
-##Check if IP exceeds warning limit
-#def exceedsWarningLimit(request, ip):
-#    currIp = get_object(ip)
-#    ret = {
-#        'bool': (currIp.count > limits.objects.get(name="WarningLimit").val)
-#    }
-#    return HttpResponse(json.dumps(ret), content_type="application/json", status=200)
-#
-##Check if IP exceeds metering limit
-#def exceedsMeteringLimit(request, ip):
-#    currIp = get_object(ip)
-#    ret = {
-#        'bool': (currIp.count > limits.objects.get(name="MeteringLimit").val)
-#    }
-#    return HttpResponse(json.dumps(ret), content_type="application/json", status=200)
